feat: 메시지 컴포넌트 분리 추가
