+++ src/pages/index.tsx
@@ -2,72 +2,17 @@ import Layout from "@/components/layout";
 import Modal, { ReserveProp } from "@/components/modal";
 import { modalState, reserveState } from "@/lib/recoilState";
 import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useRecoilState } from "recoil";
 import DetailModal from "@/components/detailModal";
 import reserveListData from "@/data/reserveList.json";
 import ReserveListItems from "@/components/reserveListItems";
-import { getMessaging, getToken, onMessage } from "firebase/messaging";
-import { initializeApp } from "firebase/app";
-
-const Index = () => {
-  const onMessageFCM = async () => {
-    //브라우저에 알림 권한을 요청한다.
-    const permission = await Notification.requestPermission();
-    if (permission !== "granted") return;
-
-    const firebaseApp = initializeApp({
-      apiKey: process.env.NEXT_PUBLIC_API_KEY,
-      authDomain: process.env.NEXT_PUBLIC_AUTH,
-      projectId: "ex-fcm-efa3f",
-      storageBucket: "ex-fcm-efa3f.appspot.com",
-      messagingSenderId: process.env.NEXT_PUBLIC_MESSAGE_SENDERID,
-      appId: process.env.NEXT_PUBLIC_APPID,
-      measurementId: process.env.NEXT_PUBLIC_MEASUREMENTID,
-    });
-    const messaging = getMessaging(firebaseApp);
-
-    getToken(messaging, {
-      vapidKey: process.env.NEXT_PUBLIC_VAPID,
-    })
-      .then((currentToken) => {
-        if (currentToken) {
-          // 정상적으로 토큰이 발급되면 콘솔에 출력한다.
-          console.log(currentToken);
-        } else {
-          console.log("예기치 않은 오류로 토큰 발급 불가");
-        }
-      })
-      .catch((err) => {
-        console.log("예기치 않은 오류 발생", err);
-      });
-  };
-  useEffect(() => {
-    onMessageFCM();
-  }, []);
-
-  return (
-    <div>
-      <h1>Hello world</h1>
-    </div>
-  );
-};
 
 export default function Home() {
   const [reservation, setReservation] = useRecoilState(modalState);
   const [reserveList, setReserveList] =
     useRecoilState<ReserveProp[]>(reserveState);
   const router = useRouter();
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setReservation(1);
-    }, 1000);
-
-    return () => {
-      clearTimeout(timer);
-    };
-  }, []);
 
   useEffect(() => {
     setReserveList(reserveListData);
